Refactor Error & not found pages
