--- frontEnd/src/components/Error/Error.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function Error() {
-  return (
-    <div />
-  );
-}
